pyproject.toml
[project]
name = "honryu"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-generativeai>=0.8.3",
    "mss>=9.0.2",
    "pillow>=11.0.0",
    "pyqt6>=6.7.1",
    "python-dotenv>=1.0.1",
    "qt-material>=2.14",
]
